full FourDNet working with bs 8

+++ model/make_model.py
@@ -394,6 +394,7 @@ class build_SimpleDepthNet(nn.Module):
         return cls_score, x
 
 
+
 class build_FourDNet(nn.Module):
     def __init__(self, num_classes, camera_num, view_num, cfg, factory, rearrange):
         print(f"<===================== building FourDNet =========================>")
@@ -416,31 +417,38 @@ class build_FourDNet(nn.Module):
         else:
             view_num = 0
 
+
+        # defining the RGB backbone
         self.base = factory[cfg.MODEL.TRANSFORMER_TYPE](img_size=cfg.INPUT.SIZE_TRAIN, sie_xishu=cfg.MODEL.SIE_COE, local_feature=cfg.MODEL.JPM, camera=camera_num, view=view_num, stride_size=cfg.MODEL.STRIDE_SIZE, drop_path_rate=cfg.MODEL.DROP_PATH)
 
         if pretrain_choice == 'imagenet':
             self.base.load_param(model_path)
             print('Loading pretrained ImageNet model......from {}'.format(model_path))
         
-        # reduce dimensionality of ViT features
+
+        # model's channel dimension
         self.reduced_dim = 128
+
+
+        # reduce dimensionality of ViT features
         self.reduce_dims = nn.Linear(self.in_planes, self.reduced_dim)
         self.reduce_dims_global = nn.Linear(self.in_planes, self.reduced_dim)
 
-        # project depth input to higher dimension
-        self.project_depth = nn.Sequential(
-            nn.Conv2d(1, self.reduced_dim // 2, 3, 2, 1),
-            nn.Conv2d(self.reduced_dim // 2, self.reduced_dim, 3, 2, 1),
-            nn.Conv2d(self.reduced_dim, self.reduced_dim, 3, 2, 1),
-            # nn.Conv2d(self.reduced_dim, self.reduced_dim, 3, 2, 1),
-        )
-        self.global_depth_feat = nn.Parameter(torch.randn(self.reduced_dim))
+        
+        # defining the D backbone 
+        self.vgg = VGGFeatures()
+        # 128 are the local features, 512 are the global features
+        self.merge_local_global_depth = nn.Linear(128 + 512, self.reduced_dim)
+
+        # the depth-classifier at the end of depth branch
+        # self.ffn = nn.Conv2d(512, self.reduced_dim, 7, 1, 0)
+        # self.classifier = nn.Linear(self.reduced_dim, num_classes)
+
 
         # R2D cross attention
         self.r2d_k = 3
         self.r2d_m = 8
         self.r2d_Q = nn.Linear(2 * self.reduced_dim, self.reduced_dim)
-        self.merge_local_global_depth = nn.Linear(2 * self.reduced_dim, self.reduced_dim)
         self.r2d_V = nn.Linear(self.reduced_dim, self.reduced_dim)
         self.r2d_selector = nn.Sequential(
             nn.Linear(self.reduced_dim, 2 * self.r2d_m * self.r2d_k),
@@ -454,21 +462,8 @@ class build_FourDNet(nn.Module):
         self.device = "cuda:0"
         self.classifier = nn.Linear(self.reduced_dim, num_classes)
 
-        self.d2d_k = 15 
-        self.d2d_m = 2
-        self.d2d_Q = nn.Linear(self.reduced_dim, self.reduced_dim)
-        self.d2d_V = nn.Linear(self.reduced_dim, self.reduced_dim)
-        self.d2d_selector = nn.Sequential(
-            nn.Linear(self.reduced_dim, 2 * self.d2d_m * self.d2d_k),
-            nn.Sigmoid(),
-        ) 
-        self.d2d_attn_weights = nn.Sequential(
-            nn.Linear(self.reduced_dim, self.d2d_m * self.d2d_k),
-            nn.Softmax(dim=-1)
-        )
-
         # development stage
-        self.visualize = True
+        self.visualize = False
         self.vis_count = 0
         self.max_vis = 100
 
@@ -480,7 +475,7 @@ class build_FourDNet(nn.Module):
         # print()
         # depth = depth.half()
         # print(f"len(np.unique(depth)) = {len(torch.unique(depth))}")
-        if self.vis_count < self.max_vis:
+        if self.visualize and self.vis_count < self.max_vis:
             if os.path.exists(f"vis"):
                 shutil.rmtree(f"vis")
             os.mkdir(f"vis")
@@ -496,92 +491,73 @@ class build_FourDNet(nn.Module):
                 plt.savefig(f"vis/{label[batch_idx]}/depth{self.vis_count}.jpg")
                 plt.close()
                 self.vis_count += 1
+
+        depth = depth.float()
+        rgb = rgb.float()
+
+
+        # extracting the RGB features
         features = self.base(rgb, cam_label=cam_label, view_label=view_label)
         N = features.shape[1] - 1
+        # print(f"features.shape = {features.shape}")
+
         # global rgb features
         global_feat = features[:, 0]
         global_feat = self.reduce_dims_global(global_feat)
+        # print(f"global_feat.shape = {global_feat.shape}")
 
         # local rgb features
         local_feat = features[:, 1:]
         local_feat = self.reduce_dims(local_feat)
+        # print(f"local_feat.shape = {local_feat.shape}")
 
         # concatenating local and global rgb features 
         local_cat_global = torch.cat((global_feat.unsqueeze(1).repeat(1, N, 1), local_feat), -1)
+        # print(f"local_cat_global.shape = {local_cat_global.shape}")
 
-        # local depth features
-        local_depth_feat = self.project_depth(depth).permute(0, 2, 3, 1)
-        Hd, Wd = local_depth_feat.shape[-3], local_depth_feat.shape[-2]
-        local_depth_feat = local_depth_feat.reshape(B, Hd * Wd, self.reduced_dim)
+        # depth features
+        features4, features20 = self.vgg(depth)
+        # features4 = torch.rand(B, 128, 112, 112).to(0)
+        # features20 = torch.rand(B, 512, 7, 7).to(0)
+        Hd, Wd = features4.shape[-2], features4.shape[-1]
+        local_depth_feat = features4.permute(0, 2, 3, 1).reshape(B, Hd * Wd, self.reduced_dim)
+        # print(f"Hd = {Hd}")
+        # print(f"Wd = {Wd}")
+        # print(f"local_depth_feat.shape = {local_depth_feat.shape}")
+
+        # global average pooling on the global depth features
+        global_depth_feat = torch.mean(features20.reshape(B, 512, 49), dim=-1)
+        # print(f"global_depth_feat.shape = {global_depth_feat.shape}")
 
         # concatenating local and global depth features 
+        # print(f"Hd = {Hd}")
+        # print(f"Wd = {Wd}")
+        # print(f"global_depth_feat.shape = {global_depth_feat.shape}")
         local_cat_global_depth = torch.cat((
-            self.global_depth_feat.unsqueeze(0).unsqueeze(0).repeat(B, Hd * Wd, 1),
+            global_depth_feat.unsqueeze(1).repeat(1, Hd * Wd, 1),
             local_depth_feat
         ), -1)
         local_cat_global_depth = self.merge_local_global_depth(local_cat_global_depth)
+        assert local_cat_global_depth.shape == (B, Hd * Wd, self.reduced_dim)
         # local_cat_global_depth = local_depth_feat
 
-        """D2D Self Attention"""
-        q = self.d2d_Q(local_cat_global_depth)
-        v = self.d2d_V(local_cat_global_depth)
-
-        selector_outputs = self.d2d_selector(q)
-        attention_scores = self.d2d_attn_weights(q)
-        locations_x = selector_outputs[:, :, 0 : self.d2d_m * self.d2d_k]
-        locations_y = selector_outputs[:, :, self.d2d_m * self.d2d_k :] 
-
-        # find out nearest positions for each selected position
-        # x1, x2, y1, y2
-        nearest_pos = torch.zeros((B, Hd * Wd, self.d2d_m * self.d2d_k, 4))
-        stride_x = 1.0 / (Wd - 1)
-        stride_y = 1.0 / (Hd - 1)
-        nearest_pos[..., 0] = locations_x // stride_x
-        nearest_pos[..., 1] = torch.minimum(nearest_pos[..., 0] + 1, torch.tensor(Wd - 1))
-        nearest_pos[..., 2] = locations_y // stride_y
-        nearest_pos[..., 3] = torch.minimum(nearest_pos[..., 2] + 1, torch.tensor(Hd - 1))
-        nearest_pos = nearest_pos.type(torch.int32)
-        nearest_feat = torch.zeros((B, Hd * Wd, self.d2d_m * self.d2d_k, 4, v.shape[-1]))
-
-
-        # nearest_feat2 = torch.zeros((B, N, self.r2d_m * self.r2d_k, 4, v.shape[-1]))
-        # for batch_idx in range(B):
-        #     nearest_feat2[batch_idx, :, :, 0] = v[batch_idx, nearest_pos[batch_idx, :, :, 0] * Wd + nearest_pos[batch_idx, :, :, 3]]
-        #     nearest_feat2[batch_idx, :, :, 1] = v[batch_idx, nearest_pos[batch_idx, :, :, 1] * Wd + nearest_pos[batch_idx, :, :, 3]]
-        #     nearest_feat2[batch_idx, :, :, 2] = v[batch_idx, nearest_pos[batch_idx, :, :, 1] * Wd + nearest_pos[batch_idx, :, :, 2]]
-        #     nearest_feat2[batch_idx, :, :, 3] = v[batch_idx, nearest_pos[batch_idx, :, :, 0] * Wd + nearest_pos[batch_idx, :, :, 2]]
-        
-
-        nearest_pos = nearest_pos.to(self.device).type(torch.int64)
-        locations_x = locations_x.to(self.device)
-        locations_y = locations_y.to(self.device)
-        nearest_feat = nearest_feat.to(self.device)
-        pos0 = nearest_pos[:, :, :, 0] * Wd + nearest_pos[:, :, :, 3]
-        pos1 = nearest_pos[:, :, :, 1] * Wd + nearest_pos[:, :, :, 3]
-        pos2 = nearest_pos[:, :, :, 1] * Wd + nearest_pos[:, :, :, 2]
-        pos3 = nearest_pos[:, :, :, 0] * Wd + nearest_pos[:, :, :, 2]
-        nearest_feat[:, :, :, 0] = torch.gather(v.unsqueeze(1).expand(B, Hd * Wd, v.shape[-2], v.shape[-1]), 2, pos0.unsqueeze(-1).expand(*pos0.shape, self.reduced_dim))
-        nearest_feat[:, :, :, 1] = torch.gather(v.unsqueeze(1).expand(B, Hd * Wd, v.shape[-2], v.shape[-1]), 2, pos1.unsqueeze(-1).expand(*pos1.shape, self.reduced_dim))
-        nearest_feat[:, :, :, 2] = torch.gather(v.unsqueeze(1).expand(B, Hd * Wd, v.shape[-2], v.shape[-1]), 2, pos2.unsqueeze(-1).expand(*pos2.shape, self.reduced_dim))
-        nearest_feat[:, :, :, 3] = torch.gather(v.unsqueeze(1).expand(B, Hd * Wd, v.shape[-2], v.shape[-1]), 2, pos3.unsqueeze(-1).expand(*pos3.shape, self.reduced_dim))
-
-        # assert torch.allclose(nearest_feat, nearest_feat2.to(self.device))
-        # print(f"assertion valid!")
-
-        interpolated_feat = bilinear_interpolation(nearest_feat, nearest_pos[..., 0], nearest_pos[..., 1], nearest_pos[..., 2], nearest_pos[..., 3], locations_x * (Wd - 1), locations_y * (Hd - 1)).squeeze(-1).squeeze(-1)
-
-        # print(f"interpolated_feat.shape = {interpolated_feat.shape}")
-        d2d_feat = torch.sum(interpolated_feat * attention_scores.unsqueeze(-1), dim=-2) 
-        local_cat_global_depth = local_cat_global_depth + d2d_feat
 
         """R2D Cross Attention"""
+        # print(f"starting with R2D Cross Attention")
         q = self.r2d_Q(local_cat_global)
         v = self.r2d_V(local_cat_global_depth)
+
+        # print(f"q.shape = {q.shape}")
+        # print(f"v.shape = {v.shape}")
 
         selector_outputs = self.r2d_selector(q)
         attention_scores = self.r2d_attn_weights(q)
         locations_x = selector_outputs[:, :, 0 : self.r2d_m * self.r2d_k]
         locations_y = selector_outputs[:, :, self.r2d_m * self.r2d_k :] 
+
+        # print(f"locations_x.shape = {locations_x.shape}")
+        # print(f"locations_y.shape = {locations_y.shape}")
+        # print(f"attention_scores.shape = {attention_scores.shape}")
 
         # find out nearest positions for each selected position
         # x1, x2, y1, y2
@@ -595,6 +571,8 @@ class build_FourDNet(nn.Module):
         nearest_pos = nearest_pos.type(torch.int32)
         nearest_feat = torch.zeros((B, N, self.r2d_m * self.r2d_k, 4, v.shape[-1]))
 
+        # print(f"nearest_pos.shape = {nearest_pos.shape}")
+        # print(f"nearest_feat.shape = {nearest_feat.shape}")
 
         # nearest_feat2 = torch.zeros((B, N, self.r2d_m * self.r2d_k, 4, v.shape[-1]))
         # for batch_idx in range(B):
@@ -617,19 +595,31 @@ class build_FourDNet(nn.Module):
         nearest_feat[:, :, :, 2] = torch.gather(v.unsqueeze(1).expand(B, N, v.shape[-2], v.shape[-1]), 2, pos2.unsqueeze(-1).expand(*pos2.shape, self.reduced_dim))
         nearest_feat[:, :, :, 3] = torch.gather(v.unsqueeze(1).expand(B, N, v.shape[-2], v.shape[-1]), 2, pos3.unsqueeze(-1).expand(*pos3.shape, self.reduced_dim))
 
+        # print(f"nearest_feat.shape = {nearest_feat.shape}")
+
         # assert torch.allclose(nearest_feat, nearest_feat2.to(self.device))
         # print(f"assertion valid!")
 
         interpolated_feat = bilinear_interpolation(nearest_feat, nearest_pos[..., 0], nearest_pos[..., 1], nearest_pos[..., 2], nearest_pos[..., 3], locations_x * (Wd - 1), locations_y * (Hd - 1)).squeeze(-1).squeeze(-1)
+        # print(f"interpolated_feat.shape = {interpolated_feat.shape}")
 
+        # print(f"interpolated_feat.shape = {interpolated_feat.shape}")
+        # print(f"attention_scores.shape = {attention_scores.shape}")
         r2d_feat = torch.sum(interpolated_feat * attention_scores.unsqueeze(-1), dim=-2) 
         r2d_feat = torch.mean(r2d_feat, -2)
-        final_depth_feat = torch.mean(local_cat_global_depth, dim=-2)
-        final_embedding = final_depth_feat 
-        # final_embedding = r2d_feat 
+        # print(f"r2d_feat.shape = {r2d_feat.shape}")
+
+        # final_depth_feat = torch.mean(local_cat_global_depth, dim=-2)
+        # final_embedding = final_depth_feat 
+        final_embedding = r2d_feat 
+
         # final_embedding = global_feat 
         cls_score = self.classifier(final_embedding)
-        return cls_score, final_embedding
+        if self.training:
+            return cls_score, final_embedding
+        else:
+            return final_embedding 
+
 
 
 class build_transformer_local(nn.Module):
@@ -823,9 +813,9 @@ def make_model(cfg, num_class, camera_num, view_num):
     #     model = Backbone(num_class, cfg)
     #     print('===========building ResNet===========')
 
-    # model = build_FourDNet(num_class, camera_num, view_num, cfg, __factory_T_type, rearrange=cfg.MODEL.RE_ARRANGE)
+    model = build_FourDNet(num_class, camera_num, view_num, cfg, __factory_T_type, rearrange=cfg.MODEL.RE_ARRANGE)
     # model = build_SimpleDepthNet(num_class, camera_num, view_num, cfg, __factory_T_type, rearrange=cfg.MODEL.RE_ARRANGE)
     # model = build_DepthNet(num_class, camera_num, view_num, cfg, __factory_T_type, rearrange=cfg.MODEL.RE_ARRANGE)
-    model = build_DepthNet2(num_class, camera_num, view_num, cfg, __factory_T_type, rearrange=cfg.MODEL.RE_ARRANGE)
+    # model = build_DepthNet2(num_class, camera_num, view_num, cfg, __factory_T_type, rearrange=cfg.MODEL.RE_ARRANGE)
     # model = build_transformer_local(num_class, camera_num, view_num, cfg, __factory_T_type, rearrange=cfg.MODEL.RE_ARRANGE)
     return model
